Update README, enhance data ingestion, and improve backend API functionality
- Revised the README to clarify the requirements for the `CENSUS_API_KEY` and added instructions for multi-year data ingestion.
- Updated `ingest.py` to support optional year parameters for data ingestion, allowing users to specify the analysis year.
- Enhanced error handling in the ingestion process to provide clearer troubleshooting guidance for API key issues.
- Improved the backend API to ensure consistent data retrieval and processing for tract demographics and risk scores.
- Added new helper scripts for streamlined multi-year data ingestion, improving usability for users managing large datasets.

+++ backend/app/models/tract.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from typing import Optional
 
-from sqlalchemy import Boolean, Float, Index, String, Text
+from sqlalchemy import Boolean, Float, Index, Integer, String, Text
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from geoalchemy2 import Geometry
 
@@ -23,7 +23,7 @@ class Tract(Base):
     centroid_lon: Mapped[Optional[float]] = mapped_column(Float, nullable=True)
     median_rent: Mapped[Optional[float]] = mapped_column(Float, nullable=True)
     median_household_income: Mapped[Optional[float]] = mapped_column(Float, nullable=True)
-    population: Mapped[Optional[float]] = mapped_column(Float, nullable=True)
+    population: Mapped[Optional[int]] = mapped_column(Integer, nullable=True)
     is_institutional: Mapped[bool] = mapped_column(Boolean, nullable=False, default=False)
     geometry: Mapped[object] = mapped_column(
         Geometry(geometry_type="MULTIPOLYGON", srid=4326, spatial_index=False),
